Now cost function variables are in double instead float

+++ src/CostFunction.cpp
@@ -2,7 +2,7 @@
 // CostFunction.cpp -- implementation of the class "CostFunction".
 //
 //    This file is part of the featsel program
-//    Copyright (C) 2015  Marcelo S. Reis
+//    Copyright (C) 2017  Marcelo S. Reis
 //
 //    This program is free software: you can redistribute it and/or modify
 //    it under the terms of the GNU General Public License as published by
@@ -87,13 +87,13 @@ void CostFunction::set_max_number_of_calls_of_cost_function
 }
 
 
-float CostFunction::get_threshold ()
+double CostFunction::get_threshold ()
 {
   return threshold;
 }
 
 
-void CostFunction::set_threshold (float value)
+void CostFunction::set_threshold (double value)
 {
   has_threshold_set = true;
   threshold = value;
