+++ src/functions/Explicit.cpp
@@ -2,7 +2,7 @@
 // Explicit.cpp -- implementation of the class "Explicit".
 //
 //    This file is part of the featsel program
-//    Copyright (C) 2015  Marcelo S. Reis
+//    Copyright (C) 2017  Marcelo S. Reis
 //
 //    This program is free software: you can redistribute it and/or modify
 //    it under the terms of the GNU General Public License as published by
@@ -33,11 +33,11 @@ Explicit::~Explicit ()
 }
 
 
-float Explicit::cost (ElementSubset * X)
+double Explicit::cost (ElementSubset * X)
 {
   timeval begin, end;
   gettimeofday (& begin, NULL);
-  float cost = 0;
+  double cost = 0;
   number_of_calls_of_cost_function++;
 
   unsigned int i;
