+++ src/functions/MeanConditionalEntropy.h
@@ -53,7 +53,7 @@ protected:
 
   // Return the MCE of a given subset X.
   //
-  float calculate_MCE (ElementSubset *);
+  double calculate_MCE (ElementSubset *);
 
   // Compute the distribution of a subset X from W-operator samples,
   // storing them into as subsets of a set S, where |S| = |X|.
@@ -64,7 +64,7 @@ protected:
   // Return the conditional entropy of Y given X = x; the distribution of the
   // samples must be computed before using this function.
   //
-  float calculate_conditional_entropy (unsigned int *, float);
+  double calculate_conditional_entropy (unsigned int *, double);
 
 public:
 
@@ -78,7 +78,7 @@ public:
 
   // Returns the value of c(X), where X is a subset.
   //
-  float cost (ElementSubset *);
+  double cost (ElementSubset *);
 
 };
