+++ src/functions/PartCost.h
@@ -2,7 +2,7 @@
 // PartCost.h -- definition of the class "PartCost".
 //
 //    This file is part of the featsel program
-//    Copyright (C) 2015  Marcelo S. Reis, Gustavo Estrela
+//    Copyright (C) 2017 Gustavo Estrela
 //
 //    This program is free software: you can redistribute it and/or modify
 //    it under the terms of the GNU General Public License as published by
@@ -56,7 +56,7 @@ public:
 
   // Returns the value of c(X), where X is a subset.
   //
-  float cost (ElementSubset *);
+  double cost (ElementSubset *);
 
 };
